Need to use image explicitly tagged by version
For some reason, even "docker ps" will gleefully remember image tags
as plain text and show an outdated tag as the image, even if the
labels changed since the corresponding container started (typical with
the "latest" tag). It is no surprise that "docker service update" is
none the wiser. To prevent surprising null updates, we have to ensure
that tag names passed to "docker service" are never re-used.