Fix authentication in staging using "proxy_redirect"
Gitlab redirects to "/" which, according to the vague wording of
http://api.rubyonrails.org/classes/ActionController/Redirecting.html#method-i-redirect_to
appears to mean that the expansion of the Location: header ignores the
external_url in the configuration. This appeared to work - So long as
the Redises were shared - because the production Gitlab would set a
cookie that the staging Gitlab would be able to validate.

(Un-separating the Redises wouldn't have fully solved the problem
anyway, as one of the tests before promoting staging to production
consists of deleting the current user and trying to log in anew; and
that breaks the aforementioned compatibility between staging and
production, since the user ID changes between them.)

A nice way to fix everything correctly, is to have the nginx reverse
proxy rewrite these redirects.

* Add new Perl variable $externally_visible_url
* Use it to set up proxy_redirect correctly in nginx.conf